renamed check route to health_check
