+++ server/routes/health_check/index.js
