+++ server/routes/index.js
@@ -4,7 +4,7 @@ const path = require('path');
 const fs = require('fs');
 
 router.use('/videos', require('./videos'));
-router.use('/check', require('./check'));
+router.use('/health_check', require('./health_check'));
 
 // Handle React routing, return all requests to React app
 router.get('*', (req, res) => {
